Fix the bug found in randomised testing;
the main loop in _uuniq did not take into account that nextline
returns a view to the Input buffer when the line fits there
entirely. The line Str was inserted as-is into the hash trie and its
content changed when the Input buffer was refilled.

The simple fix in this commit just allocates all lines strings
returned by nextline in the arena. This is not optimal as it
unnecessarily copies non-unique lines to the arena to be discarded
when the duplicate is found in the hash trie. This can be improved in
a later optimastion pass on the code, in this commit it was important
to record a simple fix to the bug.